output config

+++ libs/files.php
@@ -1,6 +1,11 @@
 <?php
+
+// dossier d'output
+if( !file_exists($output_folder)){
+    mkdir($output_folder) || die('impossible de créer le dossier d\' output. ' . $output_folder);
+}
 // écrire dans les fichiers
-$out = __DIR__ . '/../output/index.html';
+$out = $output_folder .'index.html';
 if( file_exists($out)){
     unlink($out);
     $messages .= '<br/> html régénéré';
